updated sensitivity values to be floats in the profileManager as that makes more sense

+++ PMMain/Midas/MidasProfileManager/ProfileWriter.h
@@ -45,7 +45,7 @@ struct AngleAction {
     std::string type;
     Command anglePositive;
     Command angleNegative;
-    unsigned int sensitivity;
+    float sensitivity;
 };
 
 struct Hold {
